Implement unilateral IVC channel teardown
These are not safe and should probably be accompanied by some change to
Communication.Rendezvous. In particular, revoking a grant before the
partner domain has unmapped it will probably cause one or two page faults.

It would be great to detect voluntary shutdown, but cleaning up after
involuntary crashes is the most important factor for this here commit.